+++ sources/framework/core/core.js
@@ -374,28 +374,6 @@
         },
 
         /**
-         * @method getRequestParameter
-         * Returns a request parameter from query string
-         * http://javablog.info/2008/04/17/url-request-parameters-using-javascript/
-         * @param {String} name - parameter name
-         * @return {String} value for the parameter or null if not found
-         */
-        getRequestParameter: function (name) {
-            // FIXME explain regex, fix escaping
-            name = name.replace(/[\[]/, '\\\[').replace(/[\]]/, '\\\]');
-            var regexS = '[\\?&]' + name + '=([^&#]*)',
-                regex = new RegExp(regexS),
-                results = regex.exec(window.location.href),
-                ret;
-            if (results === null || results === undefined) {
-                ret = null;
-            } else {
-                ret = results[1];
-            }
-            return ret;
-        },
-
-        /**
          * @method getObjectName
          * Returns Oskari event/request name from the event/request object
          * @param {Oskari.mapframework.request.Request/Oskari.mapframework.event.Event} obj